Use unsigned int instead of size_t where possible

+++ include/fill_rand.hpp
@@ -89,7 +89,7 @@ fill_rand(T& x)
         std::unreachable();
 }
 
-template <std::unsigned_integral T, size_t N>
+template <std::unsigned_integral T, unsigned int N>
 void
 fill_rand(std::array<T, N>& container)
 {
